+++ demos/browser/index.html
@@ -241,11 +241,12 @@
 		import { doAppStart, execGenSlidesFunc, table2slides1, table2slides2 } from './browser.js';
 
 		// TODO: update all other buttons like below
+		$('#btnGenFunc_Chart').click(() => execGenSlidesFunc('Chart'));
 		$('#btnGenFunc_Image').click(() => execGenSlidesFunc('Image'));
 		$('#btnGenFunc_Media').click(() => execGenSlidesFunc('Media'));
 		$('#btnGenFunc_Shape').click(() => execGenSlidesFunc('Shape'));
-		$('#btnGenFunc_Table').click(() => execGenSlidesFunc('Table'));
 		$('#btnGenFunc_Text').click(() => execGenSlidesFunc('Text'));
+		$('#btnGenFunc_Table').click(() => execGenSlidesFunc('Table'));
 		$('#btnGenFunc_Master').click(() => execGenSlidesFunc('Master'));
 
 		$('#table2slides1').click(() => table2slides1());
@@ -855,7 +856,7 @@
 					</div>
 				</div>
 				<div class="text-center">
-					<button type="button" class="btn btn-success w-25" onclick="execGenSlidesFunc('Chart')">Generate Chart Demo</button>
+					<button id="btnGenFunc_Chart" type="button" class="btn btn-success w-25">Generate Chart Demo</button>
 				</div>
 			</div>
 			<div class="tab-pane shadow p-4" id="images" role="tabpanel" aria-labelledby="images-tab">
